requirements.txt
appdirs==1.4.4
build==1.3.0
buildozer==1.5.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
Cython==0.29.37
distlib==0.4.0
docutils==0.22.2
filelock==3.19.1
filetype==1.2.0
idna==3.10
Jinja2==3.1.6
Kivy>=2.3.0
Kivy-Garden==0.1.5
kivymd2>=2.0.0
MarkupSafe==3.0.3
packaging==25.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
ptyprocess==0.7.0
Pygments==2.19.2
pyproject_hooks==1.2.0
python-for-android==2024.1.21
requests==2.32.5
sh==1.14.3
toml==0.10.2
urllib3==2.5.0
virtualenv==20.34.0
